Issue #1836 - Moves milestones initialisations in config
This is a temporary fix. The part of the code which download the
milestones.json and save them as a file should be part of the
deployment script, which in itself should probably handle the local dev
code. See Milestone Architecture Austin.
