+++ config/__init__.py
@@ -7,9 +7,64 @@
 '''This module handles application configuration and secrets.'''
 
 from collections import namedtuple
+import json
+import os
+import sys
+import urlparse
 
-from environment import *
-from secrets import *
+import requests
+
+from environment import *  # nopep8
+from secrets import *  # nopep8
+
+
+def initialize_status():
+    """Map the status name with the milestone id on GitHub.
+
+    The project needs the mapping in between milestones and their id
+    to be able to query status related things. It uses a backup version
+    when the request to GitHub fails.
+    """
+    print('Statuses Initialization…')
+    REPO_ROOT = ISSUES_REPO_URI.rpartition('/issues')[0]
+    milestones_url_path = os.path.join('repos', REPO_ROOT, 'milestones')
+    milestones_url = urlparse.urlunparse(
+        ('https', 'api.github.com', milestones_url_path, '', '', ''))
+    milestones_path = os.path.join(DATA_PATH, 'milestones.json')
+    try:
+        # Get the milestone from the network
+        print('Fetching milestones from Github…')
+        r = requests.get(milestones_url)
+        milestones_content = r.content
+        if r.status_code == 200:
+            with open(milestones_path, 'w') as f:
+                f.write(r.content)
+            print('Milestones saved in data/')
+        r.raise_for_status()
+    except requests.exceptions.HTTPError as error:
+        # Not working, let's use the cached copy
+        # This might fail the first time.
+        print(MILESTONE_ERROR.format(msg=error))
+        with open(milestones_path, 'r') as f:
+            milestones_content = f.read()
+    finally:
+        # save in data/ the current version
+        if milestones_content:
+            app.config['STATUSES'] = convert_milestones(milestones_content)
+            app.config['JSON_STATUSES'] = json.dumps(app.config['STATUSES'])
+            print('Milestones in memory')
+            return True
+        else:
+            return False
+
+
+def convert_milestones(milestones_content):
+    """Convert the JSON milestones from GitHub to a simple dict."""
+    milestone_full = json.loads(milestones_content)
+    for milestone in milestone_full:
+        STATUSES[milestone['title']]['id'] = milestone['number']
+    return STATUSES
+
 
 THREADS_PER_PAGE = 8
 
@@ -67,3 +122,8 @@ EXTRA_LABELS = [
     'type-webvr',
     'type-stylo'
 ]
+
+from webcompat import app
+# We need the milestones
+if not initialize_status():
+    sys.exit('Milestones are not initialized.')
